Remove legacy action files
