+++ lib/steps/google/assign-saml-profile/execute.ts
@@ -1,7 +1,39 @@
 "use server";
-import type { StepContext, StepExecutionResult } from "@/lib/types";
-import { executeG7AssignGoogleSamlToRootOu } from "@/app/actions/execution-actions";
 
-export async function executeAssignSamlProfile(context: StepContext): Promise<StepExecutionResult> {
-  return executeG7AssignGoogleSamlToRootOu(context);
+import * as google from "@/lib/api/google";
+import type { StepContext, StepExecutionResult } from "@/lib/types";
+import { OUTPUT_KEYS } from "@/lib/types";
+import { portalUrls } from "@/lib/api/url-builder";
+import { getGoogleToken } from "../utils/auth";
+import { handleExecutionError } from "../utils/error-handling";
+
+/**
+ * Assign the configured SAML profile to the Root OU in Google Workspace.
+ */
+export async function executeAssignSamlProfile(
+  context: StepContext,
+): Promise<StepExecutionResult> {
+  try {
+    const token = await getGoogleToken();
+    const profileFullName = context.outputs[
+      OUTPUT_KEYS.GOOGLE_SAML_PROFILE_FULL_NAME
+    ] as string;
+    if (!profileFullName) {
+      return {
+        success: false,
+        error: { message: "Missing SAML profile name." },
+      };
+    }
+    await google.assignSamlToOrgUnits(token, profileFullName, [
+      { orgUnitId: "/", ssoMode: "SAML_SSO_ENABLED" },
+    ]);
+    return {
+      success: true,
+      message:
+        "SAML profile assigned to Root OU for all users. Specific assignments can be adjusted in Google Admin console.",
+      resourceUrl: portalUrls.google.sso.main(),
+    };
+  } catch (e) {
+    return handleExecutionError(e, "G-7");
+  }
 }
