+++ lib/steps/google/create-automation-ou/execute.ts
@@ -3,11 +3,15 @@
 import * as google from "@/lib/api/google";
 import type { StepContext, StepExecutionResult } from "@/lib/types";
 import { OUTPUT_KEYS } from "@/lib/types";
+import { portalUrls } from "@/lib/api/url-builder";
 import { getGoogleToken } from "../utils/auth";
 import { handleExecutionError } from "../utils/error-handling";
 
+/**
+ * Create the 'Automation' Organizational Unit in Google Workspace.
+ */
 export async function executeCreateAutomationOu(
-  _context: StepContext
+  _context: StepContext,
 ): Promise<StepExecutionResult> {
   try {
     const token = await getGoogleToken();
@@ -26,7 +30,7 @@ export async function executeCreateAutomationOu(
             [OUTPUT_KEYS.AUTOMATION_OU_ID]: existing.orgUnitId,
             [OUTPUT_KEYS.AUTOMATION_OU_PATH]: existing.orgUnitPath,
           },
-          resourceUrl: `https://admin.google.com/ac/orgunits/details?ouPath=${existing.orgUnitPath}`,
+          resourceUrl: portalUrls.google.orgUnits.details(existing.orgUnitPath),
         };
       }
       return {
@@ -49,7 +53,7 @@ export async function executeCreateAutomationOu(
         [OUTPUT_KEYS.AUTOMATION_OU_ID]: result.orgUnitId,
         [OUTPUT_KEYS.AUTOMATION_OU_PATH]: result.orgUnitPath,
       },
-      resourceUrl: `https://admin.google.com/ac/orgunits/details?ouPath=${result.orgUnitPath}`,
+      resourceUrl: portalUrls.google.orgUnits.details(result.orgUnitPath),
     };
   } catch (e) {
     return handleExecutionError(e, "G-1");
