+++ lib/steps/microsoft/assign-users-sso/check.ts
@@ -1,8 +1,12 @@
 "use server";
 
+/**
+ * Check if any users or groups are assigned to the SAML SSO application.
+ */
+
 import type { StepCheckResult, StepContext } from "@/lib/types";
 import { OUTPUT_KEYS } from "@/lib/types";
-import { checkMicrosoftAppAssignments } from "@/app/actions/check-actions";
+import { checkMicrosoftAppAssignments } from "../utils/common-checks";
 
 export async function checkAssignUsers(context: StepContext): Promise<StepCheckResult> {
   const spId = context.outputs[OUTPUT_KEYS.SAML_SSO_SP_OBJECT_ID] as string;
