+++ lib/steps/microsoft/authorize-provisioning/check.ts
@@ -1,8 +1,13 @@
 "use server";
 
+/**
+ * Determine whether the Azure provisioning connection has been authorized.
+ * If a provisioning job exists and reports success, the step is complete.
+ */
+
 import type { StepCheckResult, StepContext } from "@/lib/types";
 import { OUTPUT_KEYS } from "@/lib/types";
-import { checkMicrosoftProvisioningJobDetails } from "@/app/actions/check-actions";
+import { checkMicrosoftProvisioningJobDetails } from "../utils/common-checks";
 
 export async function checkAuthorizeProvisioning(context: StepContext): Promise<StepCheckResult> {
   const spId = context.outputs[OUTPUT_KEYS.PROVISIONING_SP_OBJECT_ID] as string;
